Add methods from the Rectangle in sketchlib
